parsed out column editor into smaller components
